+++ apps/newsletters-ui/src/app/components/SchemaForm/ArrayInput.tsx
